+++ src/drom_lib/commandBuildDeps.ml
@@ -8,7 +8,7 @@
 (*                                                                        *)
 (**************************************************************************)
 
-open Ezcmd.TYPES
+open Ezcmd.V2
 
 let cmd_name = "build-deps"
 
@@ -20,9 +20,8 @@ let action ~args () =
 
 let cmd =
   let args, specs = Build.build_args () in
-  { cmd_name;
-    cmd_action = (fun () -> action ~args ());
-    cmd_args = [] @ specs;
-    cmd_man = [];
-    cmd_doc = "Install build dependencies only"
-  }
+  EZCMD.sub
+    cmd_name
+    (fun () -> action ~args ())
+    ~args: specs
+    ~doc: "Install build dependencies only"