+++ src/drom_lib/commandSphinx.ml
@@ -8,13 +8,17 @@
 (*                                                                        *)
 (**************************************************************************)
 
-open Ezcmd.TYPES
+open Ezcmd.V2
+open EZCMD.TYPES
 
 let cmd_name = "sphinx"
 
 let make_sphinx p =
   let dir = Misc.sphinx_target p in
   let sphinx_target = Format.sprintf "_drom/docs/%s" dir in
+  let before_script =  "scripts/before-sphinx.sh" in
+  if Sys.file_exists before_script then
+    Misc.call [| before_script ; sphinx_target |];
   Misc.call [| "sphinx-build"; "sphinx"; sphinx_target |];
   sphinx_target
 
@@ -28,14 +32,24 @@ let action ~args ~open_www () =
 let cmd =
   let args, specs = Build.build_args () in
   let open_www = ref false in
-  { cmd_name;
-    cmd_action = (fun () -> action ~args ~open_www ());
-    cmd_args =
+  EZCMD.sub cmd_name
+    (fun () -> action ~args ~open_www ())
+    ~args: (
       [ ( [ "view" ],
           Arg.Set open_www,
-          Ezcmd.info "Open a browser on the sphinx documentation" )
+          EZCMD.info "Open a browser on the sphinx documentation" )
       ]
-      @ specs;
-    cmd_man = [];
-    cmd_doc = "Generate documentation using sphinx"
-  }
+      @ specs
+    )
+    ~doc: "Generate documentation using sphinx"
+    ~man: [
+      `S "DESCRIPTION";
+      `Blocks [
+        `P "This command performs the following actions:";
+        `I ("1.", "Build the project, installing dev dependencies if not done yet (see $(b,drom build) and $(b,drom dev-deps) for more info).");
+        `I ("2.", "If a file $(i,scripts/before-sphinx.sh) exists, run it");
+        `I ("3.", "Build Sphinx documentation using the command $(b,sphinx-build sphinx _drom/docs/${sphinx-target}), where $(b,${sphinx-target}) is the $(b,sphinx-target) field in the project description, or $(b,sphinx) by default. Documentation source files are expected to be found in the top $(b,sphinx/) directory.");
+        `I ("4.", "If the argument $(b,--view) was specified, open a browser on the newly generated documentation");
+
+      ]
+    ]