+++ src/drom_lib/commandTree.ml
@@ -8,7 +8,7 @@
 (*                                                                        *)
 (**************************************************************************)
 
-open Ezcmd.TYPES
+open Ezcmd.V2
 open Types
 
 let cmd_name = "tree"
@@ -99,10 +99,30 @@ let action () =
   print_deps "" "tools" p.tools;
   ()
 
-let cmd =
-  { cmd_name;
-    cmd_action = (fun () -> action ());
-    cmd_args = [];
-    cmd_man = [];
-    cmd_doc = "Display a tree of dependencies"
-  }
+let cmd = EZCMD.sub cmd_name action
+    ~doc: "Display a tree of dependencies"
+    ~man:[
+      `S "DESCRIPTION";
+
+      `Blocks [
+        `P "Print the project as a tree of dependencies, i.e. dependencies are printed as branches of the package they are dependencies of. If a package is itself a dependency of another package, it will be printed there.";
+      ];
+
+      `S "EXAMPLE";
+      `Pre {|
+└──drom (/src/drom)
+   └──drom_lib (/src/drom_lib)
+      └──toml 5.0.0
+      └──opam-file-format 2.1.1
+      └──ez_subst >= 0.1
+      └──ez_file 0.2.0
+      └──ez_config 0.1.0
+      └──ez_cmdliner 0.2.0
+      └──directories >= 0.2
+[tools]
+└── ppx_inline_test
+└── ppx_expect
+└── odoc
+└── ocamlformat
+|};
+    ]