+++ src/drom_lib/commandUninstall.ml
@@ -8,7 +8,7 @@
 (*                                                                        *)
 (**************************************************************************)
 
-open Ezcmd.TYPES
+open Ezcmd.V2
 
 let cmd_name = "uninstall"
 
@@ -19,9 +19,7 @@ let action ~args () =
 
 let cmd =
   let args, specs = Build.build_args () in
-  { cmd_name;
-    cmd_action = (fun () -> action ~args ());
-    cmd_args = [] @ specs;
-    cmd_man = [];
-    cmd_doc = "Uninstall the project from the project opam switch"
-  }
+  EZCMD.sub cmd_name
+    (fun () -> action ~args ())
+    ~args: specs
+    ~doc: "Uninstall the project from the project opam switch"