+++ src/drom_lib/subst.ml
@@ -8,6 +8,7 @@
 (*                                                                        *)
 (**************************************************************************)
 
+open Ez_subst.V1
 open Types
 open EzCompat
 open EzFile.OP
@@ -417,17 +418,17 @@ let subst_encode p_subst escape p s =
 let project context ?bracket p s =
   try
     let escape = ref false in
-    Ez_subst.string ~sep:'!' ~escape
+    EZ_SUBST.string ~sep:'!' ~escape
       ~brace:(subst_encode project_brace escape)
       ~paren:(subst_encode project_paren (ref true))
-      ?bracket (context, p) s
+      ?bracket ~ctxt:(context, p) s
   with ReplaceContent content -> content
 
 let package context ?bracket p s =
   try
     let escape = ref false in
-    Ez_subst.string ~sep:'!' ~escape
+    EZ_SUBST.string ~sep:'!' ~escape
       ~brace:(subst_encode package_brace escape)
       ~paren:(subst_encode package_paren (ref true))
-      ?bracket (context, p) s
+      ?bracket ~ctxt:(context, p) s
   with ReplaceContent content -> content