+++ src/drom_lib/types.ml
@@ -108,10 +108,12 @@ and profile = { flags : string StringMap.t }
 
 type config =
   { config_author : string option;
+    config_share_dir : string option;
     config_github_organization : string option;
     config_license : string option;
     config_copyright : string option;
-    config_opam_repo : string option
+    config_opam_repo : string option;
+    config_dev_tools : string list;
   }
 
 type opam_kind =